{{cookiecutter.project_slug}}/requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt

{% if cookiecutter.windows == 'y' -%}
# Python-PostgreSQL Database Adapter
# If using Win for dev, this assumes Unix in prod
# ------------------------------------------------
psycopg2==2.6.2
{%- endif %}

# WSGI Handler
# ------------------------------------------------
gevent==1.2a2
gunicorn==19.6.0

# Static and Media Storage
# ------------------------------------------------
boto==2.45.0
django-storages-redux==1.3.2
{% if cookiecutter.use_whitenoise != 'y' -%}
Collectfast==0.4.0
{%- endif %}

# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==0.6.1

{% if cookiecutter.use_sentry_for_error_reporting == "y" -%}
# Raven is the Sentry client
# --------------------------
raven==5.32.0
{%- endif %}

{% if cookiecutter.use_opbeat == "y" -%}
# Opbeat agent for performance monitoring
# -----------------------------------------
opbeat==3.5.2
{%- endif %}

# Test requirements
# -----------------
coverage==4.2
flake8==3.2.1 # pyup: != 2.6.0
django-test-plus==1.0.16
factory-boy==2.8.1

# pytest
pytest-django==3.1.2
pytest-sugar==0.7.1